+++ src/main/scala/com/jswarburton/adventofcode/day07/AmplificationCircuit.scala
@@ -5,19 +5,19 @@ import com.jswarburton.adventofcode.intcode.IntCode
 import scala.io.Source
 
 object AmplificationCircuit {
-  def maxThrusterSignal(program: List[Int]): Int = {
-    val phaseSequence = List(0, 1, 2, 3, 4)
+  def maxThrusterSignal(program: List[Long]): Long = {
+    val phaseSequence = List(0L, 1L, 2L, 3L, 4L)
     val permutations = phaseSequence.permutations
 
     permutations.map(thrusterSignal(program, _)).max
   }
 
-  def thrusterSignal(program: List[Int], phaseSequence: List[Int]): Int =
-    phaseSequence.foldLeft(0)((res, phase) => IntCode.runIntCode(program, List(phase, res)).head)
+  def thrusterSignal(program: List[Long], phaseSequence: List[Long]): Long =
+    phaseSequence.foldLeft(0L)((res, phase) => IntCode.runIntCode(program, List(phase, res)).head)
 
-  def read(filePath: String): List[Int] =
+  def read(filePath: String): List[Long] =
     Source.fromFile(filePath).getLines.toList.head.split(",")
-      .map(_.toInt)
+      .map(_.toLong)
       .toList
 
 }
